Extract test methods to another file.
